+++ shell_reverse/Makefile
@@ -1,0 +1,4 @@
+all:
+	dart compile exe bin/main.dart -o ../shell_reverse/release/main.exe
+clear:
+	rm release/main.exe
