+++ shell_reverse/README.md
@@ -1,1 +1,1 @@
-A simple command-line application.
+A simple shell reverse in dart
